bench - multiple tables

+++ benchmark.py
@@ -4,6 +4,7 @@ import zlib
 import json
 from time import time
 from random import choices,seed,shuffle,random
+from tqdm import tqdm
 
 def get_data(records, key_size, val_size, doc=False):
 	t0 = time()
@@ -59,7 +60,7 @@ def benchmark_get_many_500(tab, data, keys):
 		tab.get_many(keys[i:i+500])
 
 
-def benchmark(db, fun_list, records=10_000, key_size=10, val_size=1000, iters=5, doc=False, compress=False):
+def benchmark(db, fun_list, label='', records=10_000, key_size=10, val_size=1000, iterations=3, tables=3, doc=False, compress=False):
 	if compress:
 		db.dumps = lambda x:zlib.compress(json.dumps(x).encode('utf8'))
 		db.loads = lambda x:json.loads(zlib.decompress(x).decode('utf8'))
@@ -69,38 +70,60 @@ def benchmark(db, fun_list, records=10_000, key_size=10, val_size=1000, iters=5,
 	
 	ts = int(time())
 	keys,data = [],[]
-	for i in range(iters):
+	for i in range(iterations):
 		k,d = get_data(records, key_size, val_size, doc=doc)
 		keys += [k]
 		data += [d]
 	print()
 
-	db.drop('bench')
-	tab = db.table('bench')
+	tab = {}
+	for i in range(tables):
+		db.drop(f'bench{i}')
+		tab[i] = db.table(f'bench{i}')
+	
+	pg = tqdm(total=tables*iterations*len(fun_list))
+	rows = []
+	gz = 'gzip:Y' if compress else 'gzip:N'
 	t0 = time()
-	for i in range(iters):
-		for fun in fun_list:
-			t1 = time()
-			fun(tab,data[i],keys[i])
-			t = time()-t1
-			print(ts, records, val_size, i, f'{t:5.2f}', str(int(records/t)).rjust(7), fun.__name__.replace('benchmark_',''), sep='  ')
+	for it in range(tables):
+		for i in range(iterations):
+			for fun in fun_list:
+				t1 = time()
+				fun(tab[it], data[i], keys[i])
+				t = time()-t1
+				row = label, ts, records, val_size, tables, iterations, gz, it, i, f'{t:5.2f}', str(int(records/t)).rjust(7), fun.__name__.replace('benchmark_','')
+				rows += [row]
+				pg.update(1)
+	del pg
 	print()
-	print(f'total_size: {tab.size()} chars (compress={compress})')
+	# TODO: version
+	print(*['label','ts','records','val_size','tables','iterations','gzip','table','iter','time','rows_per_s','benchmark'])
+	rows.sort(key=lambda x:(x[-1], x[7], x[8]))
+	for row in rows:
+		print(*row, sep='  ')
+		# TODO: append to file
+	print()
+	print(f'table_size: {tab[0].size()} chars (compress={compress})')
 	print(f'total_time: {time()-t0:0.1f} seconds')
 
 if __name__=="__main__":
 	import tkv
 	import os
+
+	if 1:
+		label = 'sqlite/disk'
+		db_path = 'deleteme.sqlite'
+		try:
+			os.remove(db_path)
+		except (FileNotFoundError,OSError):
+			pass
+	else:
+		label = 'sqlite_table/mem'
+		db_path = ':memory:'
+	db = tkv.connect_table(db_path)
 	
-	
-	db_path = 'deleteme.sqlite'
-	try:
-		os.remove(db_path)
-	except FileNotFoundError:
-		pass
-	db = tkv.connect(db_path)
 	seed(42)
-	benchmark(db, 
+	benchmark(db,
 		[
 			benchmark_put,
 			benchmark_get,
@@ -113,8 +136,11 @@ if __name__=="__main__":
 			benchmark_get_many_250,
 			benchmark_get_many_500,
 		],
-		records=10_000,
-		val_size=1_000,
-		compress=False,
+		label=label,
+		records=100_000,
+		val_size=100,
+		iterations=3,
+		tables=3,
+		compress=True,
 		doc=True,
 	)
